+++ Assets/Scripts/UnityUtils.Tests/Invocation/ReliableAction/InvocationTests.cs
@@ -1,4 +1,3 @@
-using System;
 using FluentAssertions;
 using NSubstitute;
 using NUnit.Framework;
@@ -21,7 +20,7 @@ namespace Invocation.ReliableAction
 
         [Test] public void ReliableAction_Invokes()
         {
-            var reliableAction = new TestReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
+            var reliableAction = new TestsReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
 
             reliableAction.TryInvoke();
 
@@ -30,7 +29,7 @@ namespace Invocation.ReliableAction
 
         [Test] public void ReliableAction_DoInvoke_OnlyOnce()
         {
-            var reliableAction = new TestReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
+            var reliableAction = new TestsReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
 
             reliableAction.TryInvoke();
             reliableAction.TryInvoke();
@@ -40,7 +39,7 @@ namespace Invocation.ReliableAction
 
         [Test] public void ReliableAction_IsInvokedProperty_ReturnsTrue_AfterSuccessfullyInvoked()
         {
-            var reliableAction = new TestReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
+            var reliableAction = new TestsReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
 
             reliableAction.TryInvoke();
 
@@ -49,7 +48,7 @@ namespace Invocation.ReliableAction
 
         [Test] public void ReliableAction_TryInvokeMethod_ReturnsTrue_WhenSuccessfullyInvoked()
         {
-            var reliableAction = new TestReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
+            var reliableAction = new TestsReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
 
             var isInvoked = reliableAction.TryInvoke();
 
@@ -58,7 +57,7 @@ namespace Invocation.ReliableAction
 
         [Test] public void ReliableAction_DoesNotInvokes_AfterCancellation()
         {
-            var reliableAction = new TestReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
+            var reliableAction = new TestsReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
 
             reliableAction.Cancel();
             reliableAction.TryInvoke();
@@ -68,7 +67,7 @@ namespace Invocation.ReliableAction
 
         [Test] public void ReliableAction_TryInvoke_ReturnsFalse_WhenInvocationCancelled()
         {
-            var reliableAction = new TestReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
+            var reliableAction = new TestsReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
 
             reliableAction.Cancel();
             var isInvoked = reliableAction.TryInvoke();
@@ -78,7 +77,7 @@ namespace Invocation.ReliableAction
 
         [Test] public void ReliableAction_IsCancelledProperty_ReturnsTrue_WhenInvocationCancelled()
         {
-            var reliableAction = new TestReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
+            var reliableAction = new TestsReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
 
             reliableAction.Cancel();
 
@@ -87,7 +86,7 @@ namespace Invocation.ReliableAction
 
         [Test] public void ReliableAction_DoesNotInvokes_WithLockedInvocation()
         {
-            var reliableAction = new TestReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
+            var reliableAction = new TestsReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
 
             var _ = reliableAction.LockInvocation();
             reliableAction.TryInvoke();
@@ -97,7 +96,7 @@ namespace Invocation.ReliableAction
 
         [Test] public void ReliableAction_IsLockedProperty_ReturnsTrue_WithLockedInvocation()
         {
-            var reliableAction = new TestReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
+            var reliableAction = new TestsReliableAction(IncrementCallsCount, _storage, _fallbackInvoker);
 
             var _ = reliableAction.LockInvocation();
 
@@ -106,28 +105,4 @@ namespace Invocation.ReliableAction
 
         private void IncrementCallsCount() => _methodCallsCount++;
     }
-
-    internal class TestReliableAction : BaseReliableAction
-    {
-        public TestReliableAction(Action action, IReliableActionsStorage storage, IFallbackInvoker invoker, bool isFallbackInvocation = false) 
-            : base(storage, invoker, isFallbackInvocation)
-        {
-            _action = action;
-        }
-
-        private readonly Action _action;
-        public override Guid TypeGuid => new("C3B7E643-9358-4FCF-9337-9BA6403F1F11");
-
-        public override void Save(string saveKey) => throw new NotImplementedException();
-        public override void Load(string saveKey) => throw new NotImplementedException();
-        public override void DeleteSave(string saveKey) => throw new NotImplementedException();
-
-        protected override void Invoke()
-        {
-            _action.Invoke();
-        }
-    }
-
-    // TODO: Test this too
-    internal class TestFallbackInvoker { }
 }
