+++ Assets/Scripts/UnityUtils.Tests/Invocation/ReliableAction/TestsFallbackInvoker.cs
@@ -1,0 +1,5 @@
+namespace Invocation.ReliableAction
+{
+    // TODO: Test this too
+    internal class TestsFallbackInvoker { }
+}
